requirements.txt
gpiozero==2.0.1
luma.core==2.5.3
luma.oled==3.14.0
openpyxl==3.0.9
pillow==12.1.0
